* refactor RAM.c and related modules * add tests for RAM module

+++ src/modules/CPU.c
@@ -118,7 +118,7 @@ void CPU_executeProgram(int *memory, FILE *outputFile)
 			case HALT:
 				break;
 			default:
-			  IO_showMessageInvalidCommand(operationCode, instructionCounter, outputFile);
+			  IO_showMessageInvalidCommand(operationCode, (int) instructionCounter, outputFile);
 			  overflowFlag = true;
 		} //end switch operation code
